Move to next.config.mjs
